# docs/openapi.json
{
  "components": {
    "schemas": {
      "AIRecommendation": {
        "description": "AI enrichment only. ``priority`` is \"high\" | \"medium\" | \"low\".",
        "properties": {
          "action": {
            "title": "Action",
            "type": "string"
          },
          "priority": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Priority"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "required": [
          "action"
        ],
        "title": "AIRecommendation",
        "type": "object"
      },
      "AIRoadmapItem": {
        "description": "AI enrichment only. ``priority`` is 1 (most important) to 10.",
        "properties": {
          "priority": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Priority"
          },
          "reasoning": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reasoning"
          },
          "skill": {
            "title": "Skill",
            "type": "string"
          }
        },
        "required": [
          "skill"
        ],
        "title": "AIRoadmapItem",
        "type": "object"
      },
      "AdminStatsResponse": {
        "properties": {
          "total_analyses": {
            "title": "Total Analyses",
            "type": "integer"
          },
          "total_profiles": {
            "title": "Total Profiles",
            "type": "integer"
          },
          "total_users": {
            "title": "Total Users",
            "type": "integer"
          }
        },
        "required": [
          "total_users",
          "total_profiles",
          "total_analyses"
        ],
        "title": "AdminStatsResponse",
        "type": "object"
      },
      "AdminUserResponse": {
        "properties": {
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "email": {
            "title": "Email",
            "type": "string"
          },
          "full_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Full Name"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "role": {
            "$ref": "#/components/schemas/UserRole"
          }
        },
        "required": [
          "id",
          "email",
          "full_name",
          "role",
          "created_at"
        ],
        "title": "AdminUserResponse",
        "type": "object"
      },
      "AnalysisRequest": {
        "description": "Request to trigger a new profile analysis.",
        "properties": {
          "target_locations": {
            "items": {
              "type": "string"
            },
            "maxItems": 5,
            "title": "Target Locations",
            "type": "array"
          },
          "target_roles": {
            "items": {
              "type": "string"
            },
            "maxItems": 10,
            "title": "Target Roles",
            "type": "array"
          }
        },
        "title": "AnalysisRequest",
        "type": "object"
      },
      "AnalysisResponse": {
        "description": "One analysis run.\n\nDETERMINISTIC (authoritative, never written by the AI):\n  jobs_analyzed_count, market_skills, strengths, skill_gaps,\n  skill_alignment, overall_alignment_score (always equal to skill_alignment).\n\nAI-DERIVED (optional enrichment, may be null/empty when the AI is unavailable):\n  ai_*, role_alignment, education_alignment, experience_alignment.\n\nA run that finishes with zero jobs is still ``completed``: jobs_analyzed_count\nis 0, market_skills is {}, strengths and skill_gaps are [] and the scores are 0.\nA run can also be returned with ``status = failed`` (HTTP 201 on POST) and an\n``error_message``; clients must check ``status`` before rendering results.",
        "properties": {
          "ai_gaps": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ai Gaps"
          },
          "ai_recommendations": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/AIRecommendation"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ai Recommendations"
          },
          "ai_relevant_roles": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ai Relevant Roles"
          },
          "ai_roadmap": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/AIRoadmapItem"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ai Roadmap"
          },
          "ai_strengths": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ai Strengths"
          },
          "ai_summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ai Summary"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "education_alignment": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Education Alignment"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "experience_alignment": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Experience Alignment"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "jobs_analyzed_count": {
            "title": "Jobs Analyzed Count",
            "type": "integer"
          },
          "market_skills": {
            "anyOf": [
              {
                "additionalProperties": {
                  "$ref": "#/components/schemas/MarketSkillEntry"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Market Skills"
          },
          "overall_alignment_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Overall Alignment Score"
          },
          "role_alignment": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Role Alignment"
          },
          "search_queries": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Search Queries"
          },
          "skill_alignment": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Skill Alignment"
          },
          "skill_gaps": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/SkillComparison"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Skill Gaps"
          },
          "status": {
            "$ref": "#/components/schemas/AnalysisStatus"
          },
          "strengths": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/SkillComparison"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Strengths"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "status",
          "jobs_analyzed_count",
          "market_skills",
          "strengths",
          "skill_gaps",
          "overall_alignment_score",
          "skill_alignment",
          "role_alignment",
          "education_alignment",
          "experience_alignment",
          "ai_summary",
          "ai_strengths",
          "ai_gaps",
          "ai_recommendations",
          "ai_relevant_roles",
          "ai_roadmap",
          "search_queries",
          "error_message",
          "created_at",
          "updated_at"
        ],
        "title": "AnalysisResponse",
        "type": "object"
      },
      "AnalysisStatus": {
        "enum": [
          "pending",
          "searching",
          "analyzing",
          "completed",
          "failed"
        ],
        "title": "AnalysisStatus",
        "type": "string"
      },
...
      "ErrorResponse": {
        "description": "Body of every non-validation error (400, 401, 403, 404, 409, 429, 500, 503).\n\nValidation failures (422) use the standard FastAPI list form instead:\n``{\"detail\": [{\"loc\": [...], \"msg\": \"...\", \"type\": \"...\"}]}``.\nClients must treat ``detail`` as ``string | ValidationIssue[]``.",
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          }
        },
        "required": [
          "detail"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MarketSkillEntry": {
        "description": "One value of ``Analysis.market_skills`` (the dict is keyed by canonical skill).\nDeterministic: computed from the analysed job sample, never by the AI.",
        "properties": {
          "canonical_skill": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Canonical Skill"
          },
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "jobs_requiring": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Jobs Requiring"
          },
          "percentage": {
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Percentage",
            "type": "number"
          }
        },
        "required": [
          "count",
          "percentage"
        ],
        "title": "MarketSkillEntry",
        "type": "object"
      },
...
      "SkillComparison": {
        "description": "One entry of ``Analysis.strengths`` (status \"strong\") or\n``Analysis.skill_gaps`` (status \"missing\"). Deterministic.",
        "properties": {
          "canonical_skill": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Canonical Skill"
          },
          "demand_percentage": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Demand Percentage"
          },
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "jobs_requiring": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Jobs Requiring"
          },
          "market_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Market Count"
          },
          "market_percentage": {
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Market Percentage",
            "type": "number"
          },
          "priority_rank": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Priority Rank"
          },
          "skill": {
            "title": "Skill",
            "type": "string"
          },
          "status": {
            "anyOf": [
              {
                "enum": [
                  "strong",
                  "missing"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          }
        },
        "required": [
          "skill",
          "market_percentage"
        ],
        "title": "SkillComparison",
        "type": "object"
      },
...
      "UserRole": {
        "enum": [
          "user",
          "admin"
        ],
        "title": "UserRole",
        "type": "string"
      },
...
      "ValidationErrorResponse": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationIssue"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "required": [
          "detail"
        ],
        "title": "ValidationErrorResponse",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "scheme": "bearer",
        "type": "http"
      },
      "cookieAuth": {
        "in": "cookie",
        "name": "access_token",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "Career Intelligence Copilot — Analyze your profile against live market data",
    "title": "Career Copilot API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/admin/stats": {
      "get": {
        "description": "System-wide statistics. Admin only.",
        "operationId": "get_system_stats_api_admin_stats_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminStatsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing, invalid or expired credentials"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Authenticated but not allowed (e.g. admin only)"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unexpected server error (no internal details)"
          }
        },
        "security": [
          {
            "cookieAuth": []
          },
          {
            "bearerAuth": []
          }
        ],
        "summary": "Get System Stats",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/admin/users": {
      "get": {
        "description": "List all users. Admin only. NEVER returns password hashes.",
        "operationId": "list_users_api_admin_users_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AdminUserResponse"
                  },
                  "title": "Response List Users Api Admin Users Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing, invalid or expired credentials"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Authenticated but not allowed (e.g. admin only)"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unexpected server error (no internal details)"
          }
        },
        "security": [
          {
            "cookieAuth": []
          },
          {
            "bearerAuth": []
          }
        ],
        "summary": "List Users",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/analysis": {
      "get": {
        "description": "List all analyses for the authenticated user.",
        "operationId": "list_analyses_api_analysis_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AnalysisResponse"
                  },
                  "title": "Response List Analyses Api Analysis Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing, invalid or expired credentials"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unexpected server error (no internal details)"
          }
        },
        "security": [
          {
            "cookieAuth": []
          },
          {
            "bearerAuth": []
          }
        ],
        "summary": "List Analyses",
        "tags": [
          "analysis"
        ]
      },
      "post": {
        "description": "Trigger a new profile analysis.\n\nPipeline:\n1. Fetch user's profile\n2. Generate market queries\n3. Search via SerpApi (cached)\n4. Normalize & deduplicate results\n5. Extract market skills\n6. Calculate demand percentages\n7. Compare user vs. market\n8. Identify strengths & gaps\n9. AI analysis & recommendations\n10. Store and return results",
        "operationId": "create_analysis_api_analysis_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalysisRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The user has no saved profile to analyse"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing, invalid or expired credentials"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                }
              }
            },
            "description": "Request body failed validation"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unexpected server error (no internal details)"
          }
        },
        "security": [
          {
            "cookieAuth": []
          },
          {
            "bearerAuth": []
          }
        ],
        "summary": "Create Analysis",
        "tags": [
          "analysis"
        ]
      }
    },
    "/api/analysis/latest": {
      "get": {
        "description": "Get the most recent analysis for the authenticated user.",
        "operationId": "get_latest_analysis_api_analysis_latest_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing, invalid or expired credentials"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The user has never run an analysis"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unexpected server error (no internal details)"
          }
        },
        "security": [
          {
            "cookieAuth": []
          },
          {
            "bearerAuth": []
          }
        ],
        "summary": "Get Latest Analysis",
        "tags": [
          "analysis"
        ]
      }
    },
    "/api/analysis/{analysis_id}": {
      "get": {
        "description": "Get a specific analysis. Ownership is enforced — users can only see their own.",
        "operationId": "get_analysis_api_analysis__analysis_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "analysis_id",
            "required": true,
            "schema": {
              "title": "Analysis Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing, invalid or expired credentials"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "No such analysis, or it belongs to another user"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unexpected server error (no internal details)"
          }
        },
        "security": [
          {
            "cookieAuth": []
          },
          {
            "bearerAuth": []
          }
        ],
        "summary": "Get Analysis",
        "tags": [
          "analysis"
        ]
      }
    },
...
  }
}